Refactor calculators to use decimal for precise financial data calculations

+++ QuoteAnalyzer/ModeCounter/DictionaryModeCounter.cs
@@ -4,16 +4,16 @@ namespace QuoteAnalyzer.ModeCounter;
 
 internal class DictionaryModeCounter : IModeCounter
 {
-    private readonly Dictionary<double, int> _counts = new();
+    private readonly Dictionary<decimal, int> _counts = new();
 
-    public void Add(double value)
+    public void Add(decimal value)
     {
         _counts[value] = _counts.GetValueOrDefault(value, 0) + 1;
     }
 
-    public double Mode()
+    public decimal Mode()
     {
-        if (_counts.Count == 0) return 0.0;
+        if (_counts.Count == 0) return 0m;
         return _counts.Aggregate((l, r) => l.Value > r.Value ? l : r).Key;
     }
 }
