+++ QuoteAnalyzer/QuoteReceiver.cs
@@ -8,13 +8,13 @@ namespace QuoteAnalyzer;
 public sealed class QuoteReceiver : IAsyncDisposable
 {
     private readonly UdpClient _udpClient;
-    private readonly ChannelWriter<double> _writer;
+    private readonly ChannelWriter<decimal> _writer;
     
     private long _networkErrors;
     private long _parseErrors;
     private long _receivedCount;
 
-    public QuoteReceiver(string multicastIP, int port, ChannelWriter<double> writer)
+    public QuoteReceiver(string multicastIP, int port, ChannelWriter<decimal> writer)
     {
         _writer = writer ?? throw new ArgumentNullException(nameof(writer));
         _udpClient = new UdpClient();
@@ -45,7 +45,7 @@ public sealed class QuoteReceiver : IAsyncDisposable
                     Interlocked.Increment(ref _receivedCount);
 
                     var text = Encoding.UTF8.GetString(result.Buffer).Trim();
-                    if (double.TryParse(text, out var value))
+                    if (decimal.TryParse(text, out var value))
                         _writer.TryWrite(value);
                     else
                         Interlocked.Increment(ref _parseErrors);
