+++ QuoteAnalyzer/Statistics/ModeCounter/IModeCounter.cs
@@ -2,6 +2,6 @@
 
 public interface IModeCounter
 {
-    void Add(double value);
-    double Mode();
+    void Add(decimal value);
+    decimal Mode();
 }
