+++ QuoteAnalyzer/Welford.cs
@@ -2,16 +2,16 @@
 
 internal class Welford
 {
-    private double _m2;
-    private double _mean;
+    private decimal _m2;
+    private decimal _mean;
 
-    public double Mean => Count > 0 ? _mean : 0.0;
+    public decimal Mean => Count > 0 ? _mean : 0m;
 
-    public double StdDev => Count > 0 ? Math.Sqrt(_m2 / Count) : 0.0;
+    public decimal StdDev => Count > 0 ? (decimal)Math.Sqrt((double)(_m2 / Count)) : 0m;
 
     public long Count { get; private set; }
 
-    public void Push(double x)
+    public void Push(decimal x)
     {
         Count++;
         var delta = x - _mean;
